Create Authorization Simulator
I built a class that can be used to simulate authentication. The class (auth) would have 2 methods namely:
- register
- login

The register method would be used to take the name,username and password from a user and store it within our class, while the login method would be used to check for the user's credentials and if it matches before it displays a message "You are logged in as name/username".